+++ OggConverter/src/Config/Updates.cs
@@ -34,7 +34,7 @@ namespace OggConverter
         /// WW - week (ex. 18 for 18th week of year)
         /// B - build of this week
         /// </summary>
-        public const int version = 19401;
+        public const int version = 19402;
 
         static bool newUpdateReady;
         static bool newPreviewReady;