+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Preview 2.5.0.0 (21.08.2019)
+
+### Added
+
+- Song names are now stored in single XML file per folder. No more tons of .mscmm files!
+- Your saved songs will be converted to new format on start
+
 ## 2.4.1 (30.07.2019)
 
 ### Added